+++ NEWS.md
@@ -10,6 +10,7 @@
 - Remove the message when a mapbayests object was passed to plot() without augment() before. #80
 - Don't stop of no observation in data (no fix, just a test actually) #23
 - Fix minor "testthat" bugs due to upgrade of R and French translation of warnings.
+- Fix bug in legend of `plot()` due to new version of ggplot 3.3.4 #82
 
 # mapbayr 0.4.1
 - Fix bugs (dependency, backward compatibility, checks)